+++ src/Models/NewsModel.php
@@ -11,6 +11,8 @@
 
 namespace HeimrichHannot\ContaoNewsAlertBundle\Models;
 
+use Contao\Model\Collection;
+
 /**
  * Class NewsModel
  *
@@ -159,23 +161,29 @@ class NewsModel extends \Contao\NewsModel
     /**
      * Returns published news, where no newsalert was sent (or newsalert should be send again.
      * Checks if there are published news with activated newsalert, where newsalert sent is not set (to 1).
-     * @param array $arrOptions
      *
-     * @return Model\Collection|NewsModel|NewsModel[]|null
+     * @param int $limit Max number of news to return
+     * @param array $options
+     *
+     * @return Collection|NewsModel|NewsModel[]|null
      */
-    public static function findUnsentPublished(array $arrOptions = [])
+    public static function findUnsentPublished($limit = 0, array $options = [])
     {
         $t = static::$strTable;
 
         $arrColumns = ["$t.newsalert_activate = 1 AND $t.newsalert_sent = 0"];
 
-        if (isset($arrOptions['ignoreFePreview']) || !BE_USER_LOGGED_IN)
+        if (isset($options['ignoreFePreview']) || !BE_USER_LOGGED_IN)
         {
             $time         = \Date::floorToMinute();
             $arrColumns[] = "($t.start='' OR $t.start<='$time') AND ($t.stop='' OR $t.stop>'" . ($time + 60) . "') AND $t.published='1'";
         }
+        if ($limit > 0)
+        {
+            $options['limit'] = $limit;
+        }
 
-        return static::findBy($arrColumns, null, $arrOptions);
+        return static::findBy($arrColumns, null, $options);
     }
 
 }